requirements.txt
Flask
flask_sqlalchemy
pyyaml
sqlalchemy_utils
requests_oauthlib
cryptography
sqlalchemy